+++ src/components/modules/Question.vue
@@ -17,7 +17,7 @@
     </div>
     <div v-if="data.editing === $firebase.auth.currentUser.uid">
       <q-card-main>
-        <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="edit('')" />
+        <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
             <q-input v-model="data.title" float-label="Title" dark />
@@ -28,6 +28,10 @@
           <div class="col-12">
             <q-input v-model="data.text" float-label="Question" type="textarea" :max-height="100" :min-rows="1" dark />
           </div>
+          <div class="col-12">
+            <q-btn color="primary" @click.native="save(id)">Save</q-btn>
+            <q-btn color="negative" @click.native="remove(id)">Delete</q-btn>
+          </div>
         </div>
       </q-card-main>
     </div>
@@ -37,7 +41,7 @@
 <script>
 export default {
   name: 'mod-text',
-  props: [ 'id', 'data', 'edit', 'remove' ],
+  props: [ 'id', 'data', 'edit', 'save', 'close', 'remove' ],
   data () {
     return {}
   }