+++ src/components/modules/Text.vue
@@ -20,13 +20,17 @@
     </div>
     <div v-if="data.editing === $firebase.auth.currentUser.uid">
       <q-card-main>
-        <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="edit('')" />
+        <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
             <q-input v-model="data.title" float-label="Title" dark />
           </div>
           <div class="col-12">
             <editor :text.sync="data.text" />
+          </div>
+          <div class="col-12">
+            <q-btn color="primary" @click.native="save(id)">Save</q-btn>
+            <q-btn color="negative" @click.native="remove(id)">Delete</q-btn>
           </div>
         </div>
       </q-card-main>
@@ -42,7 +46,7 @@ export default {
     Editor
   },
   name: 'mod-text',
-  props: [ 'id', 'data', 'edit', 'remove' ],
+  props: [ 'id', 'data', 'edit', 'save', 'close', 'remove' ],
   data () {
     return {}
   }