+++ src/plugins/database.js
@@ -54,6 +54,24 @@ function view (type, id, callback) {
   })
 }
 
+function update (type, id, data, callback) {
+  firebase.auth().currentUser.getIdToken(/* forceRefresh */ true).then(function (idToken) {
+    axios.post('/update', {
+      type: type,
+      id: id,
+      data: data,
+      token: idToken
+    })
+      .then((res) => {
+        console.log(res.data)
+        callback(res.data)
+      })
+      .catch((err) => {
+        console.log(err)
+      })
+  })
+}
+
 function search (type, terms, options, callback) {
   firebase.auth().currentUser.getIdToken(/* forceRefresh */ true).then(function (idToken) {
     axios.post('/search', {
@@ -134,6 +152,7 @@ export default ({ app, router, Vue }) => {
     add: add,
     list: list,
     view: view,
+    update: update,
     search: search,
     bible: bible,
     resources: resources,